+++ CHANGELOG.md
@@ -15,6 +15,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Add empty root types automatically when extending them https://github.com/nuwave/lighthouse/pull/1347
 - Configure a default `guard` for all authentication functionality https://github.com/nuwave/lighthouse/pull/1343
 - Configure the default amount of items in paginated lists with `pagination.default_count` https://github.com/nuwave/lighthouse/pull/1352
+- Add new methods `has()`, `overwrite()` and `registerNew()` to `TypeRegistry` to control if types should
+  be overwritten when registering duplicates
 
 ### Changed
 
@@ -26,6 +28,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Deprecated
 
 - The setting `paginate_max_count` will change to `pagination.max_count` https://github.com/nuwave/lighthouse/pull/1352
+- The `registerNew()` method of `TypeRegistry` will be removed in favor of `register()`, which will change
+  its behavior to throw when registering duplicates
 
 ## 4.12.4